feat(backend): production hardening — pooling, graceful shutdown, rate-limit & data-integrity fixes
Reliability:
- Tune the async SQLAlchemy pool (size/overflow/timeout/recycle/pre_ping) instead of defaults.
- Dispose the engine on shutdown for graceful connection draining.
- Harden the boto3 S3 client with connect/read timeouts and bounded retries.

Security:
- Rate limiter no longer grants the higher 'authenticated' tier on mere header
  presence: only a cryptographically valid JWT earns it; API-key-only and
  forged/expired tokens key by IP, closing the per-credential bucket-rotation bypass.
- Parse X-Forwarded-For honouring trusted_proxy_count (right-most trusted hop)
  so clients can't spoof their rate-limit identity.
- Verify the JWT issuer (iss) claim on every decode.
- Tighten CORS to explicit methods/headers and add a request body-size ceiling (413).

Data integrity:
- Atomic total_versions increment (SET col = col + 1) instead of read-modify-write.
- Delete the just-uploaded artifact if the version write fails (no storage/DB drift).
- Store released_at as timezone-aware UTC.
- Unique index on api_keys.key_hash (also speeds the per-request auth lookup) and
  widen reference_corpora.size_tokens to BigInteger (migration 0006).

API:
- Map FastAPI 422 validation errors into the standard ErrorEnvelope shape.
- Rate-limit the public artifact download endpoint.

Verified: ruff + mypy(strict) clean, 171 pytest pass, migration 0006 up/down against Postgres 16.

+++ backend/app/models/contamination.py
@@ -4,7 +4,7 @@ from datetime import datetime
 from typing import TYPE_CHECKING, Any
 from uuid import UUID, uuid4
 
-from sqlalchemy import Boolean, DateTime, Float, ForeignKey, Integer, Text, func
+from sqlalchemy import BigInteger, Boolean, DateTime, Float, ForeignKey, Integer, Text, func
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.dialects.postgresql import UUID as PGUUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
@@ -22,7 +22,7 @@ class ReferenceCorpus(Base):
     name: Mapped[str] = mapped_column(Text, nullable=False)
     description: Mapped[str | None] = mapped_column(Text)
     version: Mapped[str | None] = mapped_column(Text)
-    size_tokens: Mapped[int | None] = mapped_column(Integer)
+    size_tokens: Mapped[int | None] = mapped_column(BigInteger)
     source_url: Mapped[str | None] = mapped_column(Text)
     minhash_index_path: Mapped[str | None] = mapped_column(Text)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
